+++ contracts/test/ChainlinkFeedStub.sol
@@ -1,0 +1,76 @@
+//SPDX-License-Identifier: MIT
+pragma solidity =0.8.13;
+
+interface AggregatorV3Interface {
+    function decimals() external view returns (uint8);
+
+    function description() external view returns (string memory);
+
+    function version() external view returns (uint256);
+
+    // getRoundData and latestRoundData should both raise "No data present"
+    // if they do not have data to report, instead of returning unset values
+    // which could be misinterpreted as actual reported values.
+    function getRoundData(
+        uint80 _roundId
+    )
+        external
+        view
+        returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
+
+    function latestRoundData()
+        external
+        view
+        returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
+}
+
+contract ChainlinkFeedStub is AggregatorV3Interface {
+    uint8 public override decimals;
+    string public override description;
+    uint256 public override version;
+
+    uint80 public roundId;
+    int256 public answer;
+    uint256 public startedAt;
+    uint256 public updatedAt;
+    uint80 public answeredInRound;
+
+    constructor(uint8 decimals_, string memory description_, uint256 version_) {
+        decimals = decimals_;
+        description = description_;
+        version = version_;
+        roundId = 0;
+    }
+
+    function setAnswer(int256 answer_) external {
+        answer = answer_;
+    }
+
+    function setRoundData(
+        uint80 roundId_,
+        int256 answer_,
+        uint256 startedAt_,
+        uint256 updatedAt_,
+        uint80 answeredInRound_
+    ) external {
+        roundId = roundId_;
+        answer = answer_;
+        startedAt = startedAt_;
+        updatedAt = updatedAt_;
+        answeredInRound = answeredInRound_;
+    }
+
+    function setRoundDataNow(int256 answer_) external {
+        answer = answer_;
+        startedAt = updatedAt = block.timestamp;
+        answeredInRound = ++roundId;
+    }
+
+    function getRoundData(uint80) external view override returns (uint80, int256, uint256, uint256, uint80) {
+        revert("NOT_IMPLEMENTED");
+    }
+
+    function latestRoundData() external view override returns (uint80, int256, uint256, uint256, uint80) {
+        return (roundId, answer, startedAt, updatedAt, answeredInRound);
+    }
+}